+++ packages/admin/dashboard/src/components/layout/pages/single-column-page/index.ts
@@ -1,1 +1,2 @@
 export * from "./single-column-page"
+export * from "./single-column-layout-component"