+++ packages/admin/dashboard/src/routes/product-options/product-option-detail/index.ts
@@ -1,0 +1,3 @@
+export { ProductOptionBreadcrumb as Breadcrumb } from "./breadcrumb"
+export { ProductOptionDetail as Component } from "./product-option-detail"
+export { productOptionLoader as loader } from "./loader"