+++ packages/admin/dashboard/src/routes/sales-channels/sales-channel-detail/index.ts
@@ -1,3 +1,3 @@
-export { SalesChannelDetailBreadcrumb as Breadcrumb } from "./breadcrumb"
+export { SalesChannelDetailBreadcrumb as Breadcrumb, seo } from "./breadcrumb"
 export { salesChannelLoader as loader } from "./loader"
 export { SalesChannelDetail as Component } from "./sales-channel-detail"