+++ www/packages/tags/src/tags/product.ts
@@ -32,6 +32,10 @@ export const product = [
     "path": "https://docs.medusajs.com/user-guide/products/import"
   },
   {
+    "title": "Manage Product Options",
+    "path": "https://docs.medusajs.com/user-guide/products/options"
+  },
+  {
     "title": "Products Overview",
     "path": "https://docs.medusajs.com/user-guide/products"
   },
@@ -58,6 +62,10 @@ export const product = [
   {
     "title": "Extend Product",
     "path": "https://docs.medusajs.com/resources/commerce-modules/product/extend"
+  },
+  {
+    "title": "Filter Products",
+    "path": "https://docs.medusajs.com/resources/commerce-modules/product/guides/filter-products"
   },
   {
     "title": "Get Variant Prices",
@@ -170,6 +178,10 @@ export const product = [
   {
     "title": "Show Products in Storefront",
     "path": "https://docs.medusajs.com/resources/storefront-development/products/list"
+  },
+  {
+    "title": "Use Product Options in Storefront",
+    "path": "https://docs.medusajs.com/resources/storefront-development/products/options"
   },
   {
     "title": "Example: Show Sale Price",